Standardisation and cleanup
Renamed all projects to remove SDK prefix.
Renamed the Blank templates to Template.
Updated libraries for PBKDF2, ZlibStream, Lacewing (zlib) and PortAudio.
Smart VS Property Sheets added. These will add defines, libraries, Unicode, etc, all based on project configuration name. They do require manual adding into the VCXPROJ file with a text editor, but this can be done in seconds.
RC file's content (the DLL properties) were made useful via VS Property Sheet.
Standardised tabbing and newlines.
Fixed clear macros bug with StringReplacer.
Standardised REFLAG usages in Handle() and Display().
Removed use of using namespace std in most projects. We're professionals.
Removed Klystrack Unicode builds. It still uses TCHAR, but base library only uses char.
Fixed Solution configurations to build the correct project configs.
Set default startup project to DarkEdif Template.
Checked JSON and tabbing in files. Updated gitattributes to compensate.
Dropped unused PortAudio library files.
Removed unnecessary text document, renamed conversion instructions.
Removed _MAC defines in RC files and headers, which are going nowhere. If necessary, I'll re-add when iOS compatibility is introduced.
Preprocessor definition cleanup:
Removed IN_EXT_VERSION2, _WINDOWS from DarkEDIF projects.
Removed _WINSOCK_DEPRECATED_NO_WARNINGS from Bluewing Server (no effect).
Switched manual defined UNICODE/_UNICODE to using Character Set = Unicode instead.
Fixed scanf_s call in Array Upgrade Ext.
Removed RC files' DISCARDABLE, PURE etc, as they are unused in Win32 and apply only to Win16.
ActiveVoxel was cleaned up and made accessible.
RangeHandler, PortAudio and MyCharSet had Runtime Unicode, but no Edittime/Debug Unicode. Fixed that.
Remove MultiThreading as a default. It was an okay idea, but too awkward.
WIN32 definition can be removed. _WIN32 is the standard.
_UNICODE as a manual definition can be removed.
Clean up all spacing in preprocessor definitions.
_DEBUG was removed as preprocessor setting; it's defined by using a Debug Runtime Library
Disabled Minimal Rebuild (deprecated VS option)
Function-level linking removed (implied by  Edit And Continue debug)
Fixed JSON and a few other places to use more-secure functions.
Added check for unchanged template JSON data in DarkEDIF.
Fixed up the resource compiling to be smarter.
Moved *.def file into the Resources project folder.
Cleaned up intermediate folders and files generated to a standardised.
Updated old SDK usages of WIN32 to use _WIN32, built into VS.
Updated the Extension.cpp comments about REFLAG.
Also reordered the Inc files into alphabetical inside VCXPROJ files.
Removed the pragma comment mmfs2.lib where needed.

+++ rSDK/ActiveVoxel/Data.h
@@ -32,8 +32,8 @@ typedef struct tagRDATA
 	//voxel_t			voxel;
 	//int				frame;
 	matrix_t		lMatrix;
-	double			sinus[3];
-	double			cosine[3];
+	float			sinus[3];
+	float			cosine[3];
 	short *			surfaceDepthGrid;
 	char *			screenDot;
 	bool*			screenRender;
